+++ test/lightning_web/live/workflow_live/edit_test.exs
@@ -2108,7 +2108,7 @@ defmodule LightningWeb.WorkflowLive.EditTest do
       assert high_priority_view
              |> has_element?("#inspector-workflow-version", "latest")
 
-      wait constant_backoff(100) |> expiry(1_000) do
+      wait constant_backoff(100) |> expiry(4_000) do
         refute low_priority_view
                |> has_element?("#inspector-workflow-version", "latest")